+++ frontend/src/hooks/usePickerScanner.tsx
